### 2022-05-31 #4445 @editorialbot
[PRE REVIEW]: Efficiently Collecting Relative Similarity Judgments from Humans
**Submitting author:** @stsievert (<a href="http://orcid.org/0000-0002-4275-3452">Scott Sievert</a>)
**Repository:** https://github.com/stsievert/salmon/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
