### 2022-06-23 a2a0ac2
find all packages but exclude tests (#134)
