### 2022-06-27 #4517 @editorialbot
[REVIEW]: Efficiently Collecting Relative Similarity Judgments from Humans
**Submitting author:** @stsievert (<a href="http://orcid.org/0000-0002-4275-3452">Scott Sievert</a>)
**Repository:** https://github.com/stsievert/salmon/
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.2

@hoechenberger,@jorgedch, & @stain, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
