### 2022-07-07 stsievert/salmon#135 @jorgedch
Fails to run local installation
### Functionality
https://github.com/openjournals/joss-reviews/issues/4517

- [ ] **Installation:**

1. Fails to run local installation command `docker-compose build` on Mac OS 12.4:

```
#0 110.6   Attempting uninstall: pyyaml
#0 110.6     Found existing installation: PyYAML 5.1.2
#0 110.6 Pip subprocess error:
#0 110.6 ERROR: Cannot uninstall 'PyYAML'. It is a distutils installed project and thus we cannot accurately determine which files belong to it which would lead to only a partial uninstall.
#0 110.6
#0 110.6
#0 110.6 failed
#0 110.6
#0 110.6 CondaEnvException: Pip failed
#0 110.6
------
failed to solve: executor failed running [/bin/sh -c conda env update -n base --file /salmon/salmon.yml --prune]: exit code: 1
```
