### 2022-07-07 stsievert/salmon#136 @jorgedch
Fails to run unit tests
### Documentation
https://github.com/openjournals/joss-reviews/issues/4517

1. Fails to run unit tests:

```
(venv) ➜  salmon git:(master) pytest

========== test session starts ==========
platform darwin -- Python 3.9.12, pytest-7.1.2, pluggy-1.0.0 -- /Users/jorge/Desktop/JOSS/salmon/venv/bin/python
cachedir: .pytest_cache
rootdir: /Users/jorge/Desktop/JOSS/salmon, configfile: setup.cfg
collected 0 items / 14 errors

...

========== short test summary info ==========
ERROR salmon/triplets/test_config.py
ERROR salmon/triplets/samplers/adaptive/search/tests/test_gram_utils.py
ERROR salmon/triplets/samplers/adaptive/search/tests/test_search_refactor.py
ERROR salmon/triplets/samplers/adaptive/search/tests/test_uncertainty.py
ERROR salmon/triplets/samplers/tests/test_rr.py
ERROR tests/test_active.py
ERROR tests/test_active_search.py
ERROR tests/test_allowabe_targets.py
ERROR tests/test_backend.py
ERROR tests/test_basic.py
ERROR tests/test_examples.py
ERROR tests/test_offline.py
ERROR tests/test_passive.py
ERROR tests/test_validation.py
```
