### 2022-07-07 stsievert/salmon#137 @jorgedch
Typo
### Software paper
https://github.com/openjournals/joss-reviews/issues/4517

- [ ] **Quality of writing:**

1. Typo:
    - Change from: "The number of human responses required will change like scale like"
    - To: "The number of human responses required will scale like"
