### 2022-07-07 stsievert/salmon#138 @jorgedch
Incorrect DOI
### Software paper
https://github.com/openjournals/joss-reviews/issues/4517

- [ ] **References:**

1. DOI given for paper "How to Model Implicit Knowledge? Similarity Learning Methods to Assess Perceptions of Visual Representations." is incorrect (10.1111/cogs.12744). It points to another paper ("Cognitive Task Analysis for Implicit Knowledge About Visual Representations With Similarity Learning Methods").
