### 2022-07-07 #4517 @jorgedch
@stsievert

### Functionality

- [ ] **Installation:**

1. https://github.com/stsievert/salmon/issues/135

### Documentation

1. https://github.com/stsievert/salmon/issues/136

### Software paper

- [ ] **Quality of writing:**

1. https://github.com/stsievert/salmon/issues/137

- [ ] **References:**

1. https://github.com/stsievert/salmon/issues/138
