+++ paper/paper.bib
@@ -126,12 +126,13 @@
 }
 
 @article{chem,
-  title={How to Model Implicit Knowledge? Similarity Learning Methods to Assess Perceptions of Visual Representations.},
-  author={Rau, Martina A and Mason, Blake and Nowak, Robert},
-  journal={International Educational Data Mining Society},
-  year={2016},
-  publisher={ERIC},
+  title={Cognitive Task Analysis for Implicit Knowledge About Visual Representations With Similarity Learning Methods},
+  author={Mason, Blake and Rau, Martina A and Nowak, Robert},
+  journal={Cognitive science},
+  year={2019},
   doi={10.1111/cogs.12744},
+  volume={43},
+  issue={9},
 }
 
 @inproceedings{next2,
