### 2022-07-08 ced3cce
better conda env in dockerfile?
