+++ Dockerfile
@@ -5,17 +5,17 @@ RUN apt-get install -y gcc cmake g++
 RUN conda -V
 
 COPY salmon.yml /salmon/salmon.yml
-RUN conda env update -n base --file /salmon/salmon.yml --prune
+RUN conda env create -f /salmon/salmon.yml
 
 VOLUME /salmon
 VOLUME /data
 COPY *.py *.cfg *.yml *.txt *.sh /salmon/
 COPY ./salmon/ /salmon/salmon/
 RUN ls /salmon
-RUN pip install -e /salmon
+RUN conda run -n salmon pip install -e /salmon
 
 RUN chmod +x /salmon/launch.sh
 RUN chmod +rw /salmon
 # ENTRYPOINT bash launch.sh
 WORKDIR /salmon
-CMD ["bash", "launch.sh"]
+CMD ["conda", "run", "-n", "salmon", "/bin/bash", "launch.sh"]
