### 2022-07-08 stsievert/salmon#135 @stsievert
Hm... I can't reproduce.

<summary>Reproduction process</summary>

The command `conda env create -f salmon.yml` completed successfully for me with `conda 4.10.3`, the version [specified in the Dockerfile][1].

Building the container also worked for me (with Docker Desktop on macOS 10.14). I ran these commands:

``` shell
git clone https://github.com/stsievert/salmon.git
cd salmon
docker-compose build
docker-compose up -d --remove-orphans
```

[1]:https://github.com/stsievert/salmon/blob/534ee6b6eb29cb81805a39a814ef55465198fca5/Dockerfile#L1

I've deployed a fix in ced3cce in #139 (which is also the latest release). Do you want to try that commit and see if it works?
