### 2022-07-08 stsievert/salmon#135 @stsievert
Perhaps a better question... why are you trying to build the Docker containers locally? Email me at joss@stsievert.com with the email address you'd like me to use with Amazon AWS; with that, I can enable running through the install process at https://docs.stsievert.com/salmon/installation.html#experimentalist.
