### 2022-07-08 #4517 @stsievert
Salmon's documentation mentions an install process on Amazon AWS: https://docs.stsievert.com/salmon/installation.html#experimentalist. I've reached out to the reviewers (@hoechenberger, @stain, @jorgedch) to provide an Amazon AWS account so they can more easily verify that process works and because Amazon EC2 machines have a non-zero cost.

@ajstewartlang let me know if that violates any ethics rules; I can revoke their permissions at any time. The relevant AMI is public (see Amazon AWS Dashboard screenshot below), so the fact I created their accounts should not provide privileged access not available to the public.
