### 2022-07-08 abecd39
make test_actives_queries_generated narrower
