+++ tests/test_active.py
@@ -77,7 +77,7 @@ def test_active_bad_keys(server, logs):
             )
 
 
-@pytest.mark.parametrize("sampler", ["ARR", "CKL"])
+@pytest.mark.parametrize("sampler", ["ARR", "Random"])
 def test_active_queries_generated(server, sampler, logs):
     # R=1 chosen because that determines when active sampling starts; this
     # test is designed to make sure no unexpected errors are thrown in
@@ -92,37 +92,27 @@ def test_active_queries_generated(server, sampler, logs):
     with logs:
         server.authorize()
         server.post("/init_exp", data={"exp": config})
-        n_active_queries = 0
+        active_queries_generated = False
         for k in range(10 * n + 1):
             q = server.get("/query").json()
-
-            ans = random.choice([q["left"], q["right"]])
-            ans = {"winner": ans, "puid": "foo", **q}
-            print(q)
-            server.post("/answer", json=ans)
-
-            if q["score"] != -9999:
-                # scored queries have been posted to the database
-                # now, only thing to test is popping off database
-                n_active_queries += 1
-            if n_active_queries == n:
-                sleep(1)
+            query = "random" if q["score"] == -9999 else "active"
+            if query == "active":
+                active_queries = True
                 break
 
             sleep(200e-3)
+
+            ans = random.choice([q["left"], q["right"]])
+            ans = {"winner": ans, "puid": "foo", **q}
+            server.post("/answer", json=ans)
+
             if k % n == 0:
                 sleep(1)
 
-    d = server.get("/responses").json()
-
-    df = pd.DataFrame(d)
-    random_queries = df["score"] == -9999
-    active_queries = ~random_queries
-    assert active_queries.sum()
-    assert random_queries.sum()
-
-    samplers = set(df.sampler.unique())
-    assert samplers == {sampler}
+    if sampler == "Random":
+        assert not active_queries_generated
+    else:
+        assert active_queries_generated
 
 
 def test_active_basics(server, logs):
