### 2022-07-08 stsievert/salmon#136 @stsievert
Hm... can't reproduce, at least when I'm running from the `salmon` environment. I think this a setup issue like "ModuleNotFound" errors (because these tests are run every commit and generally pass). What tracebacks are the failing tests providing?

``` shell
$ conda activate salmon
(salmon) $ pytest salmon
============================ test session starts =============================
platform darwin -- Python 3.8.13, pytest-7.1.2, pluggy-1.0.0 -- /Users/scott/anaconda3/envs/salmon/bin/python
cachedir: .pytest_cache
rootdir: /Users/scott/Downloads/salmon, configfile: setup.cfg
plugins: anyio-3.5.0
collected 57 items

salmon/triplets/test_config.py::test_old_html_keys_warns[debrief] PASSED [  1%]
salmon/triplets/test_config.py::test_old_html_keys_warns[instructions] PASSED [  3%]
# ...
```
