### 2022-07-08 stsievert/salmon#136 @jorgedch
Hi @stsievert, yes you are right. My mistake for not having installed the requirements beforehand. Once I install them, I get the following (looking better):

```
(venv) ➜  salmon git:(master)

========== short test summary info ==========
ERROR tests/test_active.py::test_samplers_per_user - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_active.py::test_active_wrong_proportion - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_active.py::test_active_bad_keys - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_active.py::test_active_queries_generated[ARR] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_active.py::test_active_queries_generated[CKL] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_active.py::test_active_basics - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets[RoundRobin] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets[Random] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_too_large[RoundRobin] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_too_large[Random] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_wrong_type[RoundRobin] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_wrong_type[Random] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_not_integers[RoundRobin] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_not_integers[Random] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_too_small[RoundRobin] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_allowabe_targets.py::test_targets_too_small[Random] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_backend.py::test_backend_basics - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_backend.py::test_init_errors_propogate - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_backend.py::test_run_errors_logged - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_basics - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_bad_file_upload - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_no_repeats - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_meta - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_saves_state - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_download_restore - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_logs - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_zip_upload - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_get_config - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_config_defaults_update - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_config_misplaced_error - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_no_init_twice - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_auth_repeated_entries - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_random_error - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_html_defaults_rendered - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_basic.py::test_defaults_acceptable_config - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_defaults_config - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_basic_examples[customize.yaml] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_basic_examples[complete.yaml] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_basic_examples[basic.yaml] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_basic_examples[default.yaml] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_directory_examples[alien-eggs] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_directory_examples[colors] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_directory_examples[zappos] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_examples.py::test_directory_examples[faces] - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_passive.py::test_init_w_queries_valid_idx - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_passive.py::test_validation_sampling - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_passive.py::test_round_robin - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_passive.py::test_round_robin_per_user - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_validation.py::test_init_w_queries_valid_idx - httpx.ConnectError: [Errno 61] Connection refused
ERROR tests/test_validation.py::test_validation_sampling - httpx.ConnectError: [Errno 61] Connection refused
========== 65 passed, 7 warnings, 50 errors in 29.43s ==========
```

Basically I was trying to run the tests from a local install as the docker build is not completing in my machine (https://github.com/stsievert/salmon/issues/135), but if I understand correctly it will only work once the container is launched and connecting to the resources created during deployment.

I'm going through these as they are part of the JOSS checklist: `Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?`. Let's solve first issue #135 and then probably the tests will be passing for me as well. Otherwise, do you have some guidance of how can I get them passing without Docker?
