### 2022-07-08 stsievert/salmon#135 @jorgedch
Yes, I run the commands that you mention (so it is already the latest release), but it was getting stuck in the `docker-compose build` step:

```
git clone https://github.com/stsievert/salmon.git
cd salmon
docker-compose build
```

I've been running this process as it is included in the [installation](https://docs.stsievert.com/salmon/installation.html#local-machine) docs, but you are right, this step is indicated as relevant for developers and not for experimentalists. That being said, after pulling your latest changes in the `review-comments` branch (https://github.com/stsievert/salmon/pull/139) it installs and runs correctly, so from all sides we can close this issue :)
