### 2022-07-08 stsievert/salmon#136 @stsievert
All the tests in `tests/` require Docker and a Salmon instance at `localhost:8421`. All the tests in `salmon/` do not require Docker.

I think the fact that #135 is resolved will resolve this issue; let me know if this issue is persisting.
