### 2022-07-30 stsievert/salmon#140 @jorgedch
Missing import
### Software paper
https://github.com/openjournals/joss-reviews/issues/4517

- [ ] **Functionality:**

1. The library `ipywidgets` is needed in `requirements.txt` to run the code in [Generating embeddings offline](https://docs.stsievert.com/salmon/offline.html#generate-embeddings)
