### 2022-07-31 stsievert/salmon#140 @stsievert
Do you have a more detailed traceback? I'm not seeing any mention of `ipywidgets` in the repo (both in #139 and on the master branch):

``` shell
(review-comments) $ pwd
/Users/scott/Developer/stsievert/salmon
(review-comments) $ grep -Hnri "ipywidgets" .
(review-comments) $ git checkout master
(master) $ grep -Hnri "ipywidgets" .
(master) $
```
