### 2022-08-02 stsievert/salmon#140 @jorgedch
Indeed, it's not going to show in the repo as it is a missing requirement in the packages indicated in `requirements.txt`. I got an error when running the example code in [Generating embeddings offline](https://docs.stsievert.com/salmon/offline.html#generate-embeddings) and it was solved after I added `ipywidgets` in `requirements.txt`. So the problem is fixed, it's only matter of adding the package to the list.

**Note:** I also recommend you to sort the packages in `requirements.txt` in alphabetical order, it saves time as the project evolves.
