### 2022-08-23 #4517 @jorgedch
@stsievert the solutions to the following issues are already posted in their corresponding threads:

Functionality:
https://github.com/stsievert/salmon/issues/140
https://github.com/stsievert/salmon/issues/141

Feel free to copy/paste the code in the relevant files and close the issues. Other than that, everything else is approved from my side and the code is ready for publication. All the best, Jorge
