### 2022-09-14 #4517 @hoechenberger
Hello @stsievert @ajstewartlang,

I am happy with everything except with the installation procedure. I managed to get Salmon to run on an AWS instance, but local installation failed, both via Docker and via `conda`.

There is a problem when running `docker-compose build`. I'm pasting the relevant parts here:
```
#0 219.1 Installing pip dependencies: ...working... Pip subprocess error:
#0 317.7 ERROR: Cannot uninstall 'PyYAML'. It is a distutils installed project and thus we cannot accurately determine which files belong to it which would lead to only a partial uninstall.
...
#0 317.7   Attempting uninstall: pyyaml
#0 317.7     Found existing installation: PyYAML 5.1.2
#0 317.7
#0 317.7 failed
#0 317.7
#0 317.7 CondaEnvException: Pip failed
#0 317.7
------
failed to solve: executor failed running [/bin/sh -c conda env update -n base --file /salmon/salmon.yml --prune]: exit code: 1
```
And for installation from the `salmon.yaml` file via `conda`, I run into:
```
Encountered problems while solving:
  - nothing provides requested bokeh 2.0.1
```
This is on macOS 12.6 on an Apple Silicon machine.

Since the AWS installation worked flawlessly, I was still able to review functionality of the software.
