+++ requirements.txt
@@ -37,3 +37,4 @@ sphinx_rtd_theme
 pytest
 jupyter-server-proxy  # to view Dask dashboard
 autodoc_pydantic  # to show config docs
+ipywidgets  # https://github.com/stsievert/salmon/issues/140
