### 2022-09-18 b4f4373
add bokeh channel to pick up 2.0.1
