### 2022-09-18 #4517 @stsievert
@hoechenberger thanks for the review! I'm glad to see everything went smoothly. I've fixed the conda install issue I think, and released v1.0.0rc5. What version of conda do you have?

[1]:https://github.com/stsievert/salmon/pull/139/commits/2d3669036ab8800bbbc5af9c251e0a4f4f33b5a8
[2]:https://github.com/stsievert/salmon/pull/139/commits/b4f43733fa5ebff760f7f3178628587fbb21f2b3
