### 2022-10-04 #4517 @stain
Sorry for my late reviews - github emails got drowned in the wrong mailbox

Although I am not an expert on this domain I have reviewed this mainly from a Web developer aspect. The author has taken many steps to make this software usable for third-party users and I am quite happy to recommend this paper.

A few small improvements can still be done as detailed below, but I would not put them as blocker for the paper.

The only concern I have is for the documentation URL, it is self-hosted but currently excluded from Web archives.

## Usability for surveyed users

I was surprised how the endpoint for users is thrown straight into the question -- perhaps the documentation or paper can suggest to what extent these screens can/should be customized, e.g. to explain the particular crowd sourcing experiment, to add privacy statements etc.

## Pip or Conda or Docker?

It's good with many install options, but it can get confusing.

<https://docs.stsievert.com/salmon/installation> is different from <https://docs.stsievert.com/salmon/offline.html> which recommends first pip, but this fails in one of the many dependencies, `bokeh`, similar to @hoechenberger above:

```
Building wheels for collected packages: bokeh, blosc, python-multipart, tornado
  Building wheel for bokeh (setup.py) ... done
  Created wheel for bokeh: filename=bokeh-2.0.1-py3-none-any.whl size=9080019 sha256=1919f56df60581979e945e10b53524d9bcf4b2e2305aff2a28bd9d74eb14f54e
  Stored in directory: /home/stain/.cache/pip/wheels/be/b4/d8/7ce778fd6e637bea03a561223a77ba6649aff8168e3c613754
  Building wheel for blosc (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for blosc (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [84 lines of output]
      Not searching for unused variables given on the command line.
      -- The C compiler identification is unknown
      CMake Error at CMakeLists.txt:3 (ENABLE_LANGUAGE):
        No CMAKE_C_COMPILER could be found.
```

This is on Ubuntu 20.04.5 TLS on WSL2 with Python 3.10.4. I did not proceed e.g. with installing cmake as I tend to move to Conda if I need many binary dependencies.

A warning should be added to the documentation in `offline.html` that there are many Python dependencies which are not always trivial to install, and that the Conda approach should be tried first.

It may be worth later to tidy up the list of dependencies, are they all actually needed?  A big list can make it hard to use salmon with other software as there can be version mismatches.

You should declare the license in setup.py so it shows up in metadata at <https://pypi.org/project/salmon-triplets/> e.g. adding classifier `License :: OSI Approved :: MIT License`

Likewise the license is not declared on the documentation page, only on GitHub. On the documentation it is worth pointing out that dependencies will have other open source licenses.

I did not check the license of all the dependencies are compatible, but it is worth checking:

* starlette_prometheu (GPL3)

as using this library could make all your source code also GP3 rather than MIT.

## Docker install

Following the main guide <https://docs.stsievert.com/salmon/installation#local-machine> I did a Docker install.
As this is effectively a containerized Conda install, all the benefits and challenges of that apply, see below.

Docker Compose build and up took slightly longer than Conda install, but completed well.

## Conda install

Conda install worked without any trouble.

The `salmon.yml` has reasonable version locking where needed. However, as the Conda packages available may change over time, it is also useful to record what exact versions worked by the author for all dependencies.

Try adding a `salmon.yml.lock` file by doing:

    conda env export > salmon.yml.lock

## Paper

The paper is well written, but could reflect better if not just AWS install is possible, or what are the limitation of local/offline usage.

## Documentation pages

The documentation pages https://docs.stsievert.com/salmon/ are well structured and easy to follow.

The documentation links from the Paper are hosted at `stsievert.com` which is the author's personal home page. While I encourage the practice of self-hosting, this domain seems to not be [included in the Wayback machine](https://web.archive.org/web/20220000000000*/https://stsievert.com/) and therefore a reasonable question will be about longevity of these documentation links.  (I've tried adding it)

I would recommend checking why the domain is exluded from Wayback Machine archiving (e.g. [robots.txt](https://stsievert.com/robots.txt) which disallows all access). I would also link to the documentation using <https://w3id.org/> permalinks in case the domain name is lost.
