### 2022-10-11 1e2e1d0
add note about launching w/ docker
