### 2022-10-11 #4517 @stsievert
Thanks for the feedback @stain! Especially for the feedback on the paper, conda install lock file, and experiment manager documentation. I've added two documentation pages, and have some more notes in https://github.com/stsievert/salmon/pull/139.

Two more notes:

* I will spend some more time on offline install, and post here when complete.
* Being excluded from the Wayback machine is intentional and is not changing. I will explore other options for hosting the docs, including readthedocs.io and hosting a PDF/zipped HTML files on GitHub/S3. I will post my solution here once I've decided.
