+++ requirements.txt
@@ -37,3 +37,4 @@ pytest
 jupyter-server-proxy  # to view Dask dashboard
 autodoc_pydantic  # to show config docs
 ipywidgets  # https://github.com/stsievert/salmon/issues/140
+starlette_exporter
