+++ docs/source/conf.py
@@ -24,9 +24,9 @@ copyright = "2019, Scott Sievert"
 author = "Scott Sievert"
 
 # The full version, including alpha/beta/rc tags
-from salmon import __version__
-
-release = __version__.replace(".dirty", "")
+# from salmon import __version__
+#
+# release = __version__.replace(".dirty", "")
 
 html_context = {
     "display_github": True,  # Integrate GitHub
