+++ .readthedocs.yaml
@@ -23,5 +23,10 @@ sphinx:
 formats:
   - pdf
 
+# Optionally declare the Python requirements required to build your docs
+python:
+   install:
+   - numpydoc
+
 conda:
   environment: salmon.yml
