+++ .readthedocs.yaml
@@ -23,10 +23,10 @@ sphinx:
 formats:
   - pdf
 
-# Optionally declare the Python requirements required to build your docs
-python:
-   install:
-   - requirements: docs/requirements.txt
+# # Optionally declare the Python requirements required to build your docs
+# python:
+#    install:
+#    - requirements: docs/requirements.txt
 
 conda:
   environment: salmon.yml
