+++ .github/workflows/docs.yml
@@ -22,6 +22,9 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: 3.7
+    - name: Install latex
+      run: |
+        sudo apt-get install -y texmaker latexmk
     - name: Add conda to system path
       run: |
         # $CONDA is an environment variable pointing to the root of the miniconda directory
