+++ _sources/deps.rst.txt
@@ -1,0 +1,1889 @@
+.. _deps:
+
+Dependencies
+============
+
+Salmon depends on the libraries below. Below this list, the current license of
+each library at the time of Salmon v1.0 is included:
+
+* Python (`license <https://docs.python.org/3/license.html>`_).
+* NumPy (`license <https://numpy.org/doc/stable/license.html>`_).
+* SciPy (`license <https://projects.scipy.org/scipylib/license.html>`_).
+* Pandas (`license <https://github.com/pandas-dev/pandas/blob/main/LICENSE>`_).
+* Cython (`license <https://github.com/cython/cython/blob/master/LICENSE.txt>`_)
+* Scikit-learn (`license <https://github.com/scikit-learn/scikit-learn/blob/main/COPYING>`_)
+* FastAPI (`license <https://github.com/tiangolo/fastapi/blob/master/LICENSE>`_)
+* PyTorch (`license <https://github.com/pytorch/pytorch/blob/master/LICENSE>`_)
+* Skorch (`license <https://github.com/skorch-dev/skorch/blob/master/LICENSE>`_)
+* Altair (`license <https://github.com/altair-viz/altair/blob/master/LICENSE>`_)
+* Bokeh (`license <https://github.com/bokeh/bokeh/blob/branch-3.0/LICENSE.txt>`_)
+* Jinja2 (`license <https://jinja.palletsprojects.com/en/3.1.x/license/>`_)
+* Redis, ReJSON (`license <https://redis.io/docs/about/license/>`_)
+* HTTPX (`license <https://github.com/encode/httpx/blob/master/LICENSE.md>`_)
+* Dask (`license <https://github.com/dask/dask/blob/main/LICENSE.txt>`_).
+* Dask Distributed (`license <https://github.com/dask/distributed/blob/main/LICENSE.txt>`_).
+* Dask-ML (`license <https://github.com/dask/dask-ml/blob/main/LICENSE.txt>`_).
+* aiofiles (`license <https://github.com/Tinche/aiofiles/blob/main/LICENSE>`_)
+* click (`license <https://github.com/pallets/click/blob/main/LICENSE.rst>`_)
+* lz4 (`license <https://github.com/lz4/lz4/blob/dev/lib/LICENSE>`_)
+* pythn-blosc (`license <https://github.com/Blosc/python-blosc/blob/main/LICENSE.txt>`_)
+* cytoolz (`license <https://github.com/pytoolz/cytoolz/blob/master/LICENSE.txt>`_)
+* ujson (`license <https://github.com/ultrajson/ultrajson/blob/main/LICENSE.txt>`_)
+* ipykernel (`license <https://pypi.org/project/ipykernel/>`_)
+* nb_conda_kernels (`license <https://github.com/Anaconda-Platform/nb_conda_kernels/blob/master/LICENSE>`_)
+* Sphinx (`license <https://github.com/sphinx-doc/sphinx/blob/master/LICENSE>`_)
+* Sphinx RTD Theme (`license <https://github.com/readthedocs/sphinx_rtd_theme/blob/master/LICENSE>`_)
+* PyYAML (`license <https://github.com/yaml/pyyaml/blob/master/LICENSE>`_)
+* matplotlib (`license <https://github.com/matplotlib/matplotlib/blob/main/LICENSE/LICENSE>`_)
+* python-multipart (`license <https://github.com/andrew-d/python-multipart/blob/master/LICENSE.txt>`_)
+* numpydoc (`license <https://github.com/numpy/numpydoc/blob/main/LICENSE.txt>`_)
+* pytest (`license <https://github.com/pytest-dev/pytest/blob/main/LICENSE>`_)
[…]
