### 2022-12-05 60981c3
clear local install, doc links
