+++ README.md
@@ -2,6 +2,9 @@
 <a href="https://github.com/stsievert/salmon/actions/workflows/test.yml">
  <img src="https://github.com/stsievert/salmon/actions/workflows/test.yml/badge.svg?branch=master" />
 </a>
+<a href="https://github.com/stsievert/salmon/actions/workflows/test_pip.yml">
+ <img src="https://github.com/stsievert/salmon/actions/workflows/test_pip.yml/badge.svg?branch=master" />
+</a>
 <a href="https://github.com/stsievert/salmon/actions/workflows/test_offline.yml">
   <img src="https://github.com/stsievert/salmon/actions/workflows/test_offline.yml/badge.svg?branch=master" />
 </a>
@@ -13,14 +16,30 @@ Salmon is a tool for efficiently generating ordinal embeddings. It relies on
 "active" machine learning algorithms to choose the most informative queries for
 humans to answer.
 
-See the documentation for more detail: https://docs.stsievert.com/salmon/
+### Documentation
+
+This documentation is available at these locations:
+
+* **Primary source**: https://docs.stsievert.com/salmon/
+* Secondary source: as [a raw PDF][pdf] (and as a [slower loading PDF][blobpdf]).
+* Secondary source: as [zipped HTML directory][ziphtml], which requires unzipping the directory
+  then opening up `index.html`.
+
+[pdf]:https://github.com/stsievert/salmon/raw/gh-pages/salmon.pdf
+[blobpdf]:https://github.com/stsievert/salmon/blob/gh-pages/salmon.pdf
+[ziphtml]:https://github.com/stsievert/salmon/archive/refs/heads/gh-pages.zip
+
+Please [file an issue][issue] if you can not access the documentation.
+
+[issue]:https://github.com/stsievert/salmon/issues/new
 
 ### Running Salmon offline
 Visit the documentation at https://docs.stsievert.com/salmon/offline.html.
 Briefly, this should work:
 
 ``` shell
-$ conda env create -f salmon.yml
+$ cd path/to/salmon
+$ conda env create -f salmon.lock.yml
 $ conda activate salmon
 (salmon) $ pip install -e .
 ```
