+++ .github/workflows/docs.yml
@@ -1,9 +1,9 @@
 name: Documentation build
 
 on: push
-# on:
-#   release:
-#     types: [published]
+  # on:
+  #   release:
+  #     types: [published]
 
 # Only run when release published (not created or edited, etc)
 # https://docs.github.com/en/actions/reference/events-that-trigger-workflows#release
@@ -38,8 +38,6 @@ jobs:
       run: |
         cd docs
         make html
-        cd build/html && zip salmon-html-docs.zip build/html/* build/html/**/* && cd ../..
-        cd ..
     - name: Build backup docs
       run: |
         cd docs
