+++ .github/workflows/docs.yml
@@ -1,9 +1,9 @@
 name: Documentation build
 
-on: push
-  # on:
-  #   release:
-  #     types: [published]
+  # on: push
+on:
+  release:
+    types: [published]
 
 # Only run when release published (not created or edited, etc)
 # https://docs.github.com/en/actions/reference/events-that-trigger-workflows#release
