+++ .github/workflows/docs.yml
@@ -31,7 +31,7 @@ jobs:
         echo $CONDA/bin >> $GITHUB_PATH
     - name: Install dependencies
       run: |
-        conda env update --file salmon.yml --name base
+        conda env update --file salmon.lock.yml --name base
     - name: Install Salmon
       run: pip install -e .
     - name: Build docs
