+++ _sources/developers.rst.txt
@@ -1,6 +1,68 @@
 Developing algorithms
 =====================
 
+Install
+-------
+
+This process is meant for developers. To launch, first download the code.  It's
+possible to download `a ZIP file of Salmon's source`_, or if Git is installed,
+to run this command:
+
+.. _a ZIP file of Salmon's source: https://github.com/stsievert/salmon/archive/refs/heads/master.zip
+
+.. code:: shell
+
+   $ git clone https://github.com/stsievert/salmon.git
+
+Then, to launch a local version of Salmon you'll need `Docker Compose`_.
+After that dependency is intalled, run the following code:
+
+.. _install Docker: https://www.docker.com/products/docker-desktop
+.. _install Git: https://git-scm.com/downloads
+
+.. code:: shell
+
+   $ cd salmon
+   $ docker-compose build
+   $ docker-compose up
+   $ # visit http://localhost:8421/init or http://localhost:8421/docs
+
+.. _Docker Compose: https://docs.docker.com/compose/install/
+
+If you make changes to this code, run these commands:
+
+.. code:: shell
+
+	$ docker-compose stop
+	$ docker-compose build
+	$ docker-compose up
+
+If you want to log into the Docker container, execute these commands:
+
+.. code:: shell
+
+   $ docker ps  # to get list of running conatiners
+   CONTAINER ID   IMAGE             ... [more info]  ...  NAMES
+   08b96fbcc4c3   salmon_server     ... [more info]  ...  salmon_server_1
+   57cb3b7652d9   redislabs/rejson  ... [more info]  ...  salmon_redis_1
+   $ docker exec -it 08b96fbcc4c3 /bin/bash
+   (base) root@08b96fbcc4c3:/salmon# conda activate salmon
+   (salmon) root@08b96fbcc4c3:/salmon#
+
+.. note::
+
+   This is an alternative way to install Salmon's dependencies. If you create a
+   file in the Docker container in ``/salmon``, it will also be written to
+   ``/path/to/salmon`` on your local machine.
+
+If you run the command ``export SALMON_DEBUG=1``, the Salmon server will watch
+for changes in the source and re-launch as necessary. This won't be perfect,
+but it will reduce the number of times required to run ``docker-compose {stop,
+build, up}``.
+
+If you run the command ``export SALMON_NO_AUTH=1``, the Salmon server will
+not require a username/password.
+
 Basics
 ------
 
