+++ .github/workflows/test_conda.yml
@@ -1,4 +1,4 @@
-name: Tests (complete w/ conda)
+name: Tests (install w/ conda)
 
 on:
   push:
