+++ .github/workflows/test.yml
@@ -1,4 +1,4 @@
-name: Tests (complete w/ conda & salmon.lock.yml)
+name: Tests (install w/ pinned conda)
 
 on:
   push:
