+++ requirements.txt
@@ -11,8 +11,6 @@ cytoolz
 ujson
 torch # cpuonly
 pyyaml
-ipykernel  # to run this environ in jupyter (required for viz)
-# nb_conda_kernels
 altair
 seaborn
 fastapi[all]
