+++ _sources/deps.rst.txt
@@ -3,6 +3,9 @@
 Dependencies
 ============
 
+Salmon is licensed under the BSD License. Details are at
+`LICENSE.txt <https://github.com/stsievert/salmon/blob/master/LICENSE.txt>`_.
+
 Salmon depends on the libraries below. Below this list, the current license of
 each library at the time of Salmon v1.0 is included:
 
@@ -29,8 +32,6 @@ each library at the time of Salmon v1.0 is included:
 * pythn-blosc (`python-block's license <https://github.com/Blosc/python-blosc/blob/main/LICENSE.txt>`_)
 * cytoolz (`cytoolz's license <https://github.com/pytoolz/cytoolz/blob/master/LICENSE.txt>`_)
 * ujson (`ujson's license <https://github.com/ultrajson/ultrajson/blob/main/LICENSE.txt>`_)
-* ipykernel (`ipykernel's license <https://pypi.org/project/ipykernel/>`_)
-* nb_conda_kernels (`nb_conda_kernel's license <https://github.com/Anaconda-Platform/nb_conda_kernels/blob/master/LICENSE>`_)
 * Sphinx (`Sphinx's license <https://github.com/sphinx-doc/sphinx/blob/master/LICENSE>`_)
 * Sphinx RTD Theme (`Sphinx RTD Theme's license <https://github.com/readthedocs/sphinx_rtd_theme/blob/master/LICENSE>`_)
 * PyYAML (`PyYAML's license <https://github.com/yaml/pyyaml/blob/master/LICENSE>`_)
@@ -42,7 +43,6 @@ each library at the time of Salmon v1.0 is included:
 * cloudpickle (`cloudpickle's license <https://github.com/cloudpipe/cloudpickle/blob/master/LICENSE>`_)
 * jupyter-server-proxy (`jupyter-server-proxy's license <https://github.com/jupyterhub/jupyter-server-proxy/blob/main/LICENSE>`_)
 * autodoc pydantic (`autodoc's license <https://github.com/mansenfranzen/autodoc_pydantic/blob/main/LICENSE>`_)
-* seaborn (`seaborn's license <https://github.com/mwaskom/seaborn/blob/master/LICENSE.md>`_)
 * starlette_exporter (`starlette_exporter's license <https://github.com/stephenhillier/starlette_exporter/blob/master/LICENSE>`_)
 
 
@@ -1228,77 +1228,6 @@ ujson
     * Copyright (c) 1988-1993 The Regents of the University of California.
     * Copyright (c) 1994 Sun Microsystems, Inc.
 
-ipykernel
----------
-
-.. code-block:: none
-
-   This project is licensed under the terms of the Modified BSD License
-   (also known as New or Revised or 3-Clause BSD), as follows:
-
-   - Copyright (c) 2015, IPython Development Team
-
-   All rights reserved.
-
-   Redistribution and use in source and binary forms, with or without
-   modification, are permitted provided that the following conditions are met:
-
-   Redistributions of source code must retain the above copyright notice, this
-   list of conditions and the following disclaimer.
-
-   Redistributions in binary form must reproduce the above copyright notice, this
-   list of conditions and the following disclaimer in the documentation and/or
-   other materials provided with the distribution.
-
-   Neither the name of the IPython Development Team nor the names of its
[…]
