+++ salmon/triplets/samplers/adaptive/__init__.py
@@ -1,8 +1,9 @@
-from salmon.triplets.samplers.adaptive._embed import (GD, OGD, CntsLRDamper,
-                                                      Embedding, GeoDamp,
-                                                      PadaDampG)
-from salmon.triplets.samplers.adaptive._noise_models import (CKL, GNMDS, SOE,
-                                                             STE, TSTE,
-                                                             Logistic)
-from salmon.triplets.samplers.adaptive._score import (InfoGainScorer,
-                                                      UncertaintyScorer)
+from salmon.triplets.samplers.adaptive._embed import (
+    GD, OGD, CntsLRDamper, Embedding, GeoDamp, PadaDampG
+)
+from salmon.triplets.samplers.adaptive._noise_models import (
+    CKL, GNMDS, SOE, STE, TSTE, Logistic
+)
+from salmon.triplets.samplers.adaptive._score import (
+    InfoGainScorer, UncertaintyScorer
+)
