### 2022-12-11 af6d520
Move imports around for offline embeddings
