+++ salmon/triplets/samplers/_random_sampling.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Tuple
 
 import numpy as np
 
-from salmon.backend.sampler import DaskClient, Path, Sampler, root
+from salmon.backend.sampler import Sampler
 from salmon.triplets.samplers.utils import Answer, Query
 
 logger = logging.getLogger(__name__)
@@ -59,6 +59,9 @@ class Random(Sampler):
         return self, False
 
     def run(self, *args, **kwargs):
+        from rejson import Path
+        root = Path.rootPath()
+
         rj = self.redis_client()
         rj.jsonset(f"stopped-{self.ident}", Path("."), True)
         return None
