+++ salmon/backend/__init__.py
@@ -1,2 +1,5 @@
-from salmon.backend.core import app
+try:
+    from salmon.backend.core import app
+except ModuleNotFoundError:
+    pass
 from salmon.backend.sampler import Sampler
