+++ setup.cfg
@@ -1,3 +1,70 @@
+[metadata]
+name = salmon-triplets
+description = Efficient crowdsourcing for ordinal embeddings
+author = Scott Sievert
+author_email = salmon@stsievert.com
+url = https://docs.stsievert.com/salmon
+long_description = Salmon is a tool for efficiently generating ordinal
+                   embeddings. It relies on active machine learning algorithms
+                   to choose the most informative queries for humans to answer.
+license = BSD 3-clause License
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: BSD License
+    Topic :: Scientific/Engineering :: Artificial Intelligence
+
+[options]
+python_requires = >=3.8
+packages = find:
+install_requires =
+    numpy>=1.18.0
+    scipy
+    pandas>=1.0.1
+    Cython
+    scikit-learn
+    skorch>=0.8.0
+    torch
+    pydantic
+
+tests_require = pytest
+
+[options.extras_require]
+
+server =
+    fastparquet
+    bokeh==2.0.1
+    jinja2<3.1.0
+    redis==3.5.*
+    dask>=2021.02.0
+    distributed>=2021.02.0
+    dask-ml
+    lz4
+    blosc
+    cytoolz
+    ujson
+    pyyaml
+    fastapi[all]
+    rejson
+    httpx
+    aiofiles
+    pyarrow
+    matplotlib
+    gunicorn
+    python-multipart
+    jupyter-server-proxy
+    ipywidgets
+    starlette_exporter
+    pyarrow
+    cloudpickle
+    fastparquet
+
+dev =
+    autodoc_pydantic
+    sphinx>=4.0.0
+    numpydoc
+    sphinx_rtd_theme
+    pytest
+
 [versioneer]
 VCS = git
 style = pep440
