+++ .github/workflows/test.yml
@@ -44,7 +44,9 @@ jobs:
       run: |
         conda env update --file salmon.lock.yml --name base
     - name: Install Salmon
-      run: pip install -e .
+      run: |
+        pip install .
+        pip install pytest
     - name: Run tests in salmon/salmon/
       run: sudo /usr/share/miniconda/bin/pytest salmon/
     - uses: docker/setup-buildx-action@v1
