### 2022-12-11 c5d83a2
change how server launches
