+++ .github/workflows/test_pip.yml
@@ -42,7 +42,7 @@ jobs:
           sudo chown -R -H -L $USER:$USER salmon docs tests
     - name: Install dependencies
       run: |
-        pip install .
+        pip install ".[server,dev]"
     - name: Run tests in salmon/salmon/
       run: sudo /usr/share/miniconda/bin/pytest salmon/
     - uses: docker/setup-buildx-action@v1
