### 2022-12-31 b6043fc
show logs w/ conda run command
