+++ setup.cfg
@@ -57,6 +57,7 @@ server =
     pyarrow
     cloudpickle
     fastparquet
+    PyYAML
 
 dev =
     autodoc_pydantic
