+++ setup.cfg
@@ -25,6 +25,7 @@ install_requires =
     skorch>=0.8.0
     torch
     pydantic
+    PyYAML
 
 tests_require = pytest
 
@@ -57,7 +58,6 @@ server =
     pyarrow
     cloudpickle
     fastparquet
-    PyYAML
 
 dev =
     autodoc_pydantic
