### 2023-02-20 1b3c3fd
ensure endpoint timing plots show
