+++ docker-compose.yml
@@ -32,10 +32,13 @@ services:
       - "./salmon/_out:/data:rw"
     command: /bin/sh -c "chmod -v 777 /data && redis-server --loadmodule /usr/lib/redis/modules/rejson.so"
 
-    # redis-server command from https://hub.docker.com/r/redislabs/rejson/dockerfile
-    #
-    # https://github.com/docker/compose/issues/3270
-    # also https://github.com/boot2docker/boot2docker/issues/1083#issuecomment-151380687
+  prom:
+    image: prom/prometheus
+    depends_on: [redis]
+    expose:
+      - 9090
+    volumes:
+      - ./docker/prom:/etc/prometheus
 
   # redismonitor:
   #   user: root
@@ -47,18 +50,10 @@ services:
   #   logging:
   #     driver: none
 
-  prom:
-    image: prom/prometheus
-    depends_on: [redis]
-    expose:
-      - 9090
-    volumes:
-      - ./docker/prom:/etc/prometheus
-
-  #  granfna:
-  #    image: grafana/grafana
-  #    ports:
-  #      - "3000:3000"
+# redis-server command from https://hub.docker.com/r/redislabs/rejson/dockerfile
+#
+# https://github.com/docker/compose/issues/3270
+# also https://github.com/boot2docker/boot2docker/issues/1083#issuecomment-151380687
 
 # To get bash on one of these machines:
 # docker exec -it <containerIdOrName> bash
