### 2023-02-28 452ac94
Feedback from JOSS review (#139)
