+++ salmon/triplets/samplers/__init__.py
@@ -1,6 +1,6 @@
-from salmon.triplets.samplers._adaptive_runners import (ARR, CKL, GNMDS, SOE,
-                                                        SRR, STE, TSTE,
-                                                        Adaptive)
+from salmon.triplets.samplers._adaptive_runners import (
+    ARR, CKL, GNMDS, SOE, SRR, STE, TSTE, Adaptive
+)
 from salmon.triplets.samplers._random_sampling import Random
 from salmon.triplets.samplers._round_robin import RoundRobin
 from salmon.triplets.samplers._test_runner import Test
