+++ salmon/__init__.py
@@ -1,6 +1,13 @@
 from salmon._version import get_versions
-from salmon.backend import app as app_algs
-from salmon.frontend import app as app
-
-__version__ = app.version = app_algs.version = get_versions()["version"]
+__version__ = get_versions()["version"]
 del get_versions
+
+try:
+    from salmon.backend import app as app_algs
+    from salmon.frontend import app as app
+
+    app.version = app_algs.version = __version__
+except (ModuleNotFoundError, ImportError):
+    pass
+
+from salmon.triplets.offline import OfflineEmbedding
