### 2023-02-28 6eb6e00
footnote about local install
