### 2023-02-28 #4517 @stsievert
Thanks for the ping @arfon! I've resolved of the concerns mentioned above (and also some concerns a Salmon user had). I've made some edits to the paper, and they're available above.

One important note: I've changed from an MIT license to an Apache BSD license.

> "therefore a reasonable question will be about longevity of these documentation links."—@stain

I've added a PDF rending of the documentation to GitHub, and link to it from the docs main page so people are aware.

> "I did not check the license of all the dependencies are compatible, but it is worth checking:"—@stain

I did check them all, they're available at https://docs.stsievert.com/salmon/deps. And I've added that to the documentation home page too: https://docs.stsievert.com/salmon/index.html

> "The salmon.yml has reasonable version locking where needed. However, as the Conda packages available may change over time,"—@stain

Good idea, implemented. Thank you. I've also minimized the dependencies (and tried to make the install more clear too).

@hoechenberger I've resolved a bunch of issues with the local installation you mentioned in https://github.com/openjournals/joss-reviews/issues/4517#issuecomment-1246725742. Could you give it a try again?

More complete notes are at the top of https://github.com/stsievert/salmon/pull/139 (and a new PR at https://github.com/stsievert/salmon/pull/142)
