### 2023-03-11 #4517 @stsievert
I've added Profs. Robert Nowak and Timothy Rogers as authors after more closely reviewing [the authorship guidelines][1]; while they are not mentioned in the commit history; however they actively contributed to the project direction.

[1]:https://joss.readthedocs.io/en/latest/submitting.html#authorship
