### 2023-03-22 95351ba
example adaptive usage
