+++ tests/test_offline.py
@@ -1,5 +1,7 @@
 from pathlib import Path
 import yaml
+import random
+from typing import Dict
 
 import numpy as np
 import numpy.linalg as LA
@@ -11,6 +13,7 @@ from salmon.triplets.offline import OfflineEmbedding
 from salmon.triplets.samplers import TSTE
 import salmon.triplets.offline
 
+ArrayLike = np.ndarray
 
 def test_salmon_import():
     """This test makes sure that no errors are raised on import
@@ -128,6 +131,35 @@ def test_offline_names_correct():
     assert len(em) == len(config["targets"])
     assert set(em.columns) == {"x", "y", "target"}
     assert (em["target"] == config["targets"]).all()
+
+
+def _answer(q: Dict[str, int], X: ArrayLike) -> int:
+    h = X[q["head"]]
+    l = X[q["left"]]
+    r = X[q["right"]]
+    if LA.norm(h - l) < LA.norm(h - r):
+        return q["left"]
+    return q["right"]
+
+def test_offline_adaptive(n=10, d=2):
+    X = np.random.uniform(size=(n, d))
+
+    sampler = TSTE(n=n, d=d, alpha=1, R=1)
+    X_hat0 = sampler.search.embedding.copy()
+    for t in range(10):
+        queries, scores, _ = sampler.get_queries()
+        _good_queries = queries[np.argsort(scores)[-4:]]
+        good_queries = [{"head": h, "left": o1, "right": o2} for h, o1, o2 in _good_queries]
+        answers = [{"winner": _answer(q, X), **q} for q in good_queries]
+        sampler.process_answers(answers)
+
+    X_hatm1 = sampler.search.embedding
+    val_queries = np.asarray([np.random.choice(n, size=3, replace=False) for _ in range(1000)])
+    ans = np.asarray([0 if LA.norm(X[h] - X[l]) < LA.norm(X[h] - X[r]) else 1 for h, l, r in val_queries])
+    score0 = sampler.score(queries, ans, embedding=X_hat0)
+    scorem1 = sampler.score(queries, ans, embedding=X_hatm1)
+
+    assert score0 + 0.1 < scorem1
 
 
 if __name__ == "__main__":
