+++ .github/workflows/test_conda.yml
@@ -13,14 +13,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
-    - name: Setup Python 3.7
-      uses: actions/setup-python@v2
+    - name: Install conda environment
+      uses: mamba-org/provision-with-micromamba@main
       with:
-        python-version: 3.7
-    - name: Add conda to system path
-      run: |
-        # $CONDA is an environment variable pointing to the root of the miniconda directory
-        echo $CONDA/bin >> $GITHUB_PATH
+        environment-file: salmon.yml
     - name: Debug info
       run: |
           echo "Running pwd..."
@@ -28,25 +24,26 @@ jobs:
           echo "\nRunning echo $USER..."
           echo $USER
           echo "\nRunning chown..."
-          sudo chown -R -H -L $USER:$USER  /home/runner/work/salmon/
+          sudo chown -R -H -L $USER:$USER /home/runner/work/salmon/
           chown -R -H -L $USER:$USER /home/runner/work/salmon/
           echo "\nRunning chmod..."
           sudo chmod -R 777 /home/runner/work/salmon/
           echo "\nRunning `ls`..."
           ls
+          echo "\nRunning `micromamba info`..."
+          micromamba info
+          echo "\nRunning `micromamba list`..."
+          micromamba list
     - name: Prepare for docker build...
       run: |
           chmod +x launch.sh
           # chown -R -H -L $USER:$USER .
           sudo chown -R -H -L $USER:$USER .
           sudo chown -R -H -L $USER:$USER salmon docs tests
-    - name: Install dependencies
-      run: |
-        conda env update --file salmon.yml --name base
     - name: Install Salmon
       run: pip install -e .
     - name: Run tests in salmon/salmon/
-      run: sudo /usr/share/miniconda/bin/pytest salmon/
+      run: pytest salmon/
     - uses: docker/setup-buildx-action@v1
       with:
         driver: docker
@@ -61,4 +58,4 @@ jobs:
       run: |
           # sudo docker-compose logs -f &  # if debugging; shows logs
           # sudo /usr/share/miniconda/bin/pytest -s
-          sudo /usr/share/miniconda/bin/pytest
+          pytest
