+++ salmon/triplets/samplers/_adaptive_runners.py
@@ -294,6 +294,7 @@ class Adaptive(Sampler):
         """
         y_hat = self.predict(X, embedding=embedding)
         y = np.asarray(y)
+        y_hat = np.asarray(y_hat)
         return (y_hat == y).mean()
 
 
