### 2023-04-09 132bccb
remove unnecessary conda tests
