+++ .github/workflows/test_conda.yml
@@ -1,61 +1,0 @@
-name: Tests (install w/ conda)
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-
-  build:
-
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - name: Install conda environment
-      uses: mamba-org/provision-with-micromamba@main
-      with:
-        environment-file: salmon.yml
-    - name: Debug info
-      run: |
-          echo "Running pwd..."
-          pwd
-          echo "\nRunning echo $USER..."
-          echo $USER
-          echo "\nRunning chown..."
-          sudo chown -R -H -L $USER:$USER /home/runner/work/salmon/
-          chown -R -H -L $USER:$USER /home/runner/work/salmon/
-          echo "\nRunning chmod..."
-          sudo chmod -R 777 /home/runner/work/salmon/
-          echo "\nRunning `ls`..."
-          ls
-          echo "\nRunning `micromamba info`..."
-          micromamba info
-          echo "\nRunning `micromamba list`..."
-          micromamba list
-    - name: Prepare for docker build...
-      run: |
-          chmod +x launch.sh
-          # chown -R -H -L $USER:$USER .
-          sudo chown -R -H -L $USER:$USER .
-          sudo chown -R -H -L $USER:$USER salmon docs tests
-    - name: Install Salmon
-      run: pip install -e .
-    - name: Run tests in salmon/salmon/
-      run: pytest salmon/
-    - uses: docker/setup-buildx-action@v1
-      with:
-        driver: docker
-    - name: Build Salmon server w/ Docker
-      env:
-        SALMON_NO_AUTH: 1
-      run: |
-          sudo --preserve-env=SALMON_NO_AUTH docker-compose up &
-          until curl 127.0.0.1:8421 > /dev/null 2>&1; do :; done  # wait for container to start
-          sudo docker ps
-    - name: Run all tests
-      run: |
-          # sudo docker-compose logs -f &  # if debugging; shows logs
-          # sudo /usr/share/miniconda/bin/pytest -s
-          pytest
