### 2023-04-10 e390c18
mamba in github tests?
