+++ Dockerfile
@@ -17,7 +17,7 @@ RUN ls /salmon
 RUN chmod +x /salmon/launch.sh
 RUN chmod +rw /salmon
 
-RUN micromamba run -n base pip install --no-deps -e /salmon[server]
+RUN micromamba run -n base pip install -e /salmon[server]
 
 WORKDIR /salmon
 CMD ["micromamba", "run", "-n", "base", "/bin/bash", "launch.sh"]
