+++ salmon.yml
@@ -36,7 +36,7 @@ dependencies:
     - redis==3.5.*  # https://github.com/RedisJSON/redisjson-py/issues/67
     - matplotlib
     - gunicorn
-    - python-multipart  # optional dep required by gunicorn
+    - python-multipart  # optional dep required by gunicorn for HTML forms
     - numpydoc
     - pytest
     - jupyter-server-proxy  # to view Dask dashboard
