+++ .github/workflows/test.yml
@@ -46,10 +46,10 @@ jobs:
         mamba env create -p ./myenv --file salmon.lock.yml
     - name: Install Salmon
       run: |
-        pip install .
-        pip install pytest
+        mamba run -p ./myenv pip install .
+        mamba run -p ./myenv pip install pytest
     - name: Run tests in salmon/salmon/
-      run: sudo /usr/share/miniconda/bin/pytest salmon/
+      run: mamba run -p ./myenv pytest salmon/
     - uses: docker/setup-buildx-action@v1
       with:
         driver: docker
@@ -64,4 +64,4 @@ jobs:
       run: |
           # sudo docker-compose logs -f &  # if debugging; shows logs
           # sudo /usr/share/miniconda/bin/pytest -s
-          sudo /usr/share/miniconda/bin/pytest
+          mamba run -p ./myenv pytest
