+++ .github/workflows/test.yml
@@ -45,11 +45,11 @@ jobs:
     - name: Install Salmon
       shell: bash -l {0}
       run: |
-        pip install .
-        pip install pytest
+        sudo pip install .
+        sudo pip install pytest
     - name: Run tests in salmon/salmon/
       shell: bash -l {0}
-      run: pytest salmon/
+      run: sudo pytest salmon/
     - uses: docker/setup-buildx-action@v1
       with:
         driver: docker
@@ -64,4 +64,4 @@ jobs:
       shell: bash -l {0}
       run: |
         # sudo docker-compose logs -f &  # if debugging; shows logs
-        pytest
+        sudo pytest
