+++ .github/workflows/test.yml
@@ -23,10 +23,14 @@ jobs:
         echo "    which {python, pip, pytest}"
         which python
         which pip
+        echo "    sudo $(which {python, pip, pytest})"
+        sudo $(which python)
+        sudo $(which pip)
+        sudo $(which pytest)
         echo "    sudo which {python, pip, pytest}"
         sudo which python
         sudo which pip
-        which pytest
+        sudo which pytest
         echo "    Running pwd..."
         pwd
         echo "    Running echo $USER..."
@@ -39,19 +43,19 @@ jobs:
         echo "    Running `ls`..."
         ls
     - name: Prepare for docker build...
-      shell: sudo bash -l {0}
+      shell: bash -l {0}
       run: |
         chmod +x launch.sh
         # chown -R -H -L $USER:$USER .
         sudo chown -R -H -L $USER:$USER .
         sudo chown -R -H -L $USER:$USER salmon docs tests
     - name: Install Salmon
-      shell: sudo bash -l {0}
+      shell: bash -l {0}
       run: |
         sudo $(which pip) install .
         sudo $(which pip) install pytest
     - name: Run tests in salmon/salmon/
-      shell: sudo bash -l {0}
+      shell: bash -l {0}
       run: |
         sudo $(which pytest) salmon/
     - uses: docker/setup-buildx-action@v1
@@ -65,7 +69,7 @@ jobs:
         until curl 127.0.0.1:8421 > /dev/null 2>&1; do :; done  # wait for container to start
         sudo docker ps
     - name: Run all tests
-      shell: sudo bash -l {0}
+      shell: bash -l {0}
       run: |
         # sudo docker-compose logs -f &  # if debugging; shows logs
         sudo $(which pytest)
