+++ tests/test_offline.py
@@ -141,15 +141,15 @@ def _answer(q: Dict[str, int], X: ArrayLike) -> int:
         return q["left"]
     return q["right"]
 
-def test_offline_adaptive(n=10, d=2):
+def test_offline_adaptive(n=10, d=1):
     rng = np.random.RandomState(42)
     X = rng.uniform(size=(n, d))
-    val_queries = np.asarray([rng.choice(n, size=3, replace=False) for _ in range(1000)])
+    val_queries = np.asarray([rng.choice(n, size=3, replace=False) for _ in range(5000)])
     val_ans = np.asarray([0 if LA.norm(X[h] - X[l]) < LA.norm(X[h] - X[r]) else 1 for h, l, r in val_queries])
 
     sampler = TSTE(n=n, d=d, alpha=1, R=1)
     score0 = sampler.score(val_queries, val_ans)
-    for t in range(30):
+    for t in range(20):
         queries, scores, _ = sampler.get_queries()
         _good_queries = queries[np.argsort(scores)[-4:]]
         good_queries = [{"head": h, "left": o1, "right": o2} for h, o1, o2 in _good_queries]
