### 2023-04-14 fb3fde1
More JOSS feedback (#142)
* footnote about local install
* example adaptive usage
* misc edits (better tests, faster Dockerfile, ...)
