+++ tests/utils.py
@@ -110,10 +110,11 @@ def _reset(server):
 def server():
     server = Server("http://127.0.0.1:8421")
     server.authorize()
-    server = _reset(server)
+    # server = _reset(server)
     sleep(1)
     yield server
     server = _reset(server)
+    sleep(5)
 
 
 class LogError(Exception):
