### 2023-04-16 2029584
Build docs faster (#143)
+ slightly less stringent tests (no random failures)
