+++ tests/test_offline.py
@@ -151,14 +151,14 @@ def test_offline_adaptive(n=10, d=1):
     score0 = sampler.score(val_queries, val_ans)
     for t in range(20):
         queries, scores, _ = sampler.get_queries()
-        _good_queries = queries[np.argsort(scores)[-4:]]
+        _good_queries = queries[np.argsort(scores)[-5:]]
         good_queries = [{"head": h, "left": o1, "right": o2} for h, o1, o2 in _good_queries]
         answers = [{"winner": _answer(q, X), **q} for q in good_queries]
         sampler.process_answers(answers)
 
     scorem1 = sampler.score(val_queries, val_ans)
     assert 0 <= score0 <= scorem1 <= 1
-    assert score0 + 0.08 < scorem1, "Improves by at least 8% after 200 answers"
+    assert score0 + 0.05 < scorem1, "Improves by at least 5% after 100 answers (often much larger)"
 
 
 if __name__ == "__main__":
