### 2023-04-16 #4517 @stsievert
Thanks @ajstewartlang! I've gone through some trouble, and released Salmon v1.0.2 with the DOI [10.5281/zenodo.7832320][doi] (which points to the latest version; the v1.0.2 DOI is [10.5281/zenodo.7832431][doi102]).

I have checked the metadata and changed the authors (related the reason I had to do more than one release; now Github integration is turned off!).

[doi]:https://www.doi.org/10.5281/zenodo.7832320
[doi102]:https://www.doi.org/10.5281/zenodo.7832431
