### 2023-04-16 #4517 @ajstewartlang
@stsievert can you confirm which branch the final version of the paper (i.e., the final version of paper.md) is in please? The branch `review2` appears to be empty. I *think* it's now in `master` but please let me know for sure. Thanks!
