### 2023-04-16 #4517 @stsievert
Whoops! Good catch. Correct, `review2` has been merged into master in https://github.com/stsievert/salmon/pull/142
