### 2023-04-17 #4517 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
