### 2023-04-17 #4517 @oliviaguest
I think it's just a book and the bot is struggling for some reason? Anyway, it looks good in the PDF to me. Is that all good for you too @stsievert?
