+++ paper/paper.md
@@ -17,7 +17,7 @@ authors:
 affiliations:
  - name: University of Wisconsin--Madison
    index: 1
-date: 11 March 2023
+date: 09 April 2022
 bibliography: paper.bib
 draft: true
 ---
