### 2023-04-17 #4517 @stsievert
Huh, I haven't changed the DOIs in a while. Both errors are with conference articles ("Adaptively learning ..." is from ICML, "NEXT: A System ..." is from NeurIPS). Both links resolve to the expected article.

Also, I've reviewed the paper again and noticed a typo in `paper/paper.md`. I've published a new release (v1.0.3), on Zenodo at [10.5281/zenodo.7837296][1] (latest release: [10.5281/zenodo.7832320][2]).

[1]:https://doi.org/10.5281/zenodo.7837296
[2]:https://doi.org/10.5281/zenodo.7832320
